requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
djangorestframework-simplejwt>=5.3

openpyxl>=3.1.2
reportlab>=4.0.9
